context/DivisionContext.py
import json
import os
from datetime import datetime

class DivisionContext:
    def __init__(self, division_id, data_directory="division_data", log_directory="division_logs"):
        self.division_id = division_id
        self.shared_resources = {}
        self.task_statuses = {}
        self.strategy_plans = {}
        self.risk_assessments = {}
        self.capabilities = {}
        self.environmental_model = {}
        self.ethical_principles = {}
        self.data_directory = data_directory
        self.log_directory = log_directory
        os.makedirs(self.data_directory, exist_ok=True)
        os.makedirs(self.log_directory, exist_ok=True)

    def update_resource(self, key, value):
        self.shared_resources[key] = value
        self.log_change(f"Updated resource: {key} = {value}")

    def update_task_status(self, task_id, status):
        self.task_statuses[task_id] = status
        self.log_change(f"Updated task status: {task_id} = {status}")

    def update_strategy_plan(self, plan_id, plan_details):
        self.strategy_plans[plan_id] = plan_details
        self.log_change(f"Updated strategy plan: {plan_id} = {plan_details}")

    def update_risk_assessment(self, risk_id, risk_details):
        self.risk_assessments[risk_id] = risk_details
        self.log_change(f"Updated risk assessment: {risk_id} = {risk_details}")

    def update_capabilities(self, capability_id, capability_details):
        self.capabilities[capability_id] = capability_details
        self.log_change(f"Updated capabilities: {capability_id} = {capability_details}")

    def update_environmental_model(self, environmental_factor, data):
        self.environmental_model[environmental_factor] = data
        self.log_change(f"Updated environmental model: {environmental_factor} = {data}")

    def update_ethical_principles(self, principle_id, principle_details):
        self.ethical_principles[principle_id] = principle_details
        self.log_change(f"Updated ethical principles: {principle_id} = {principle_details}")

    def log_change(self, message):
        log_path = os.path.join(self.log_directory, f"{self.division_id}_changes.log")
        timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
        log_entry = f"{timestamp} - {message}\n"
        with open(log_path, 'a') as log_file:
            log_file.write(log_entry)

    def save_state(self):
        state = {
            "shared_resources": self.shared_resources,
            "task_statuses": self.task_statuses,
            "strategy_plans": self.strategy_plans,
            "risk_assessments": self.risk_assessments,
            "capabilities": self.capabilities,
            "environmental_model": self.environmental_model,
            "ethical_principles": self.ethical_principles,
        }
        path = os.path.join(self.data_directory, f"{self.division_id}_state.json")
        with open(path, 'w') as file:
            json.dump(state, file, indent=4)

    def load_state(self):
        path = os.path.join(self.data_directory, f"{self.division_id}_state.json")
        if os.path.exists(path):
            with open(path, 'r') as file:
                state = json.load(file)
                self.shared_resources = state.get("shared_resources", {})
                self.task_statuses = state.get("task_statuses", {})
                self.strategy_plans = state.get("strategy_plans", {})
                self.risk_assessments = state.get("risk_assessments", {})
                self.capabilities = state.get("capabilities", {})
                self.environmental_model = state.get("environmental_model", {})
                self.ethical_principles = state.get("ethical_principles", {})

    def serialize_to_text(self):
        text_path = os.path.join(self.log_directory, f"{self.division_id}_context.txt")
        with open(text_path, 'w') as text_file:
            for attr, value in self.__dict__.items():
                if attr not in ['division_id', 'data_directory', 'log_directory']:
                    text_file.write(f"{attr} ({datetime.now().strftime('%Y-%m-%d %H:%M:%S')}):\n{json.dumps(value, indent=4)}\n\n")

# Usage example:
division_context = DivisionContext('global_strategy')
division_context.update_resource('budget', 100000)
division_context.update_task_status('task001', 'completed')
division_context.save_state()
division_context.load_state()
division_context.serialize_to_text()
